scripts/requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
Pillow>=10.0.0
python-slugify>=8.0.0